+++ test/test_mesh_seq.py
@@ -6,17 +6,13 @@ import re
 import unittest
 
 from firedrake import (
-    Function,
-    FunctionSpace,
     UnitCubeMesh,
     UnitIntervalMesh,
     UnitSquareMesh,
 )
 from parameterized import parameterized
 
-from goalie.field import Field
 from goalie.mesh_seq import MeshSeq
-from goalie.time_partition import TimeInterval, TimePartition
 
 
 class BaseClasses:
@@ -26,13 +22,8 @@ class BaseClasses:
 
     class MeshSeqTestCase(unittest.TestCase):
         """
-        Test case with a simple setUp method and mesh constructor.
+        Test case with a simple mesh constructor.
         """
-
-        def setUp(self):
-            self.field = Field("field", family="Real")
-            self.time_partition = TimePartition(1.0, 2, [0.5, 0.5], self.field)
-            self.time_interval = TimeInterval(1.0, [0.5], self.field)
 
         def trivial_mesh(self, dim):
             try:
@@ -53,60 +44,13 @@ class TestExceptions(BaseClasses.MeshSeqTestCase):
     def test_inconsistent_dim_error(self):
         meshes = [self.trivial_mesh(2), self.trivial_mesh(3)]
         with self.assertRaises(ValueError) as cm:
-            MeshSeq(self.time_partition, meshes)
+            MeshSeq(meshes)
         msg = "Meshes must all have the same topological dimension."
         self.assertEqual(str(cm.exception), msg)
 
-    def test_get_solver_notimplemented_error(self):
-        mesh_seq = MeshSeq(self.time_interval, self.trivial_mesh(2))
-        with self.assertRaises(NotImplementedError) as cm:
-            mesh_seq.get_solver()
-        self.assertEqual(str(cm.exception), "'get_solver' needs implementing.")
-
-    def test_return_dict_error(self):
-        with self.assertRaises(AssertionError) as cm:
-            MeshSeq(
-                self.time_interval,
-                self.trivial_mesh(2),
-                get_initial_condition=lambda _: 0,
-            )
-        msg = "get_initial_condition should return a dict"
-        self.assertEqual(str(cm.exception), msg)
-
-    def test_missing_field_error(self):
-        with self.assertRaises(AssertionError) as cm:
-            MeshSeq(
-                self.time_interval,
-                self.trivial_mesh(2),
-                get_initial_condition=lambda _: {},
-            )
-        msg = "missing fields {'field'} in get_initial_condition"
-        self.assertEqual(str(cm.exception), msg)
-
-    def test_unexpected_field_error(self):
-        with self.assertRaises(AssertionError) as cm:
-            MeshSeq(
-                self.time_interval,
-                self.trivial_mesh(2),
-                get_initial_condition=lambda _: {"field": None, "extra_field": None},
-            )
-        msg = "unexpected fields {'extra_field'} in get_initial_condition"
-        self.assertEqual(str(cm.exception), msg)
-
-    def test_solver_generator_error(self):
-        mesh = self.trivial_mesh(2)
-        f_space = FunctionSpace(mesh, "CG", 1)
-        kwargs = {
-            "get_initial_condition": lambda _: {"field": Function(f_space)},
-            "get_solver": lambda _: lambda *_: {},
-        }
-        with self.assertRaises(AssertionError) as cm:
-            MeshSeq(self.time_interval, mesh, **kwargs)
-        self.assertEqual(str(cm.exception), "solver should yield")
-
     @parameterized.expand([1, 3])
     def test_plot_dim_error(self, dim):
-        mesh_seq = MeshSeq(self.time_interval, self.trivial_mesh(dim))
+        mesh_seq = MeshSeq(self.trivial_mesh(dim))
         with self.assertRaises(ValueError) as cm:
             mesh_seq.plot()
         self.assertEqual(str(cm.exception), "MeshSeq plotting only supported in 2D.")
@@ -120,18 +64,18 @@ class TestGeneric(BaseClasses.MeshSeqTestCase):
     def test_setitem(self):
         mesh1 = UnitSquareMesh(1, 1, diagonal="left")
         mesh2 = UnitSquareMesh(1, 1, diagonal="right")
-        mesh_seq = MeshSeq(self.time_interval, [mesh1])
+        mesh_seq = MeshSeq([mesh1])
         self.assertEqual(mesh_seq[0], mesh1)
         mesh_seq[0] = mesh2
         self.assertEqual(mesh_seq[0], mesh2)
 
     def test_counting_2d(self):
-        mesh_seq = MeshSeq(self.time_interval, [UnitSquareMesh(3, 3)])
+        mesh_seq = MeshSeq([UnitSquareMesh(3, 3)])
         self.assertEqual(mesh_seq.count_elements(), [18])
         self.assertEqual(mesh_seq.count_vertices(), [16])
 
     def test_counting_3d(self):
-        mesh_seq = MeshSeq(self.time_interval, [UnitCubeMesh(3, 3, 3)])
+        mesh_seq = MeshSeq([UnitCubeMesh(3, 3, 3)])
         self.assertEqual(mesh_seq.count_elements(), [162])
         self.assertEqual(mesh_seq.count_vertices(), [64])
 
@@ -142,34 +86,34 @@ class TestStringFormatting(BaseClasses.MeshSeqTestCase):
     Goalie's :class:`MeshSeq` object.
     """
 
-    def test_mesh_seq_time_interval_str(self):
-        mesh_seq = MeshSeq(self.time_interval, [UnitSquareMesh(1, 1)])
+    def test_mesh_seq_str(self):
+        mesh_seq = MeshSeq([UnitSquareMesh(1, 1)])
         got = re.sub("#[0-9]*", "?", str(mesh_seq))
         self.assertEqual(got, "['<Mesh ?>']")
 
-    def test_mesh_seq_time_partition_str(self):
+    def test_mesh_seq_2_str(self):
         meshes = [
             UnitSquareMesh(1, 1, diagonal="left"),
             UnitSquareMesh(1, 1, diagonal="right"),
         ]
-        mesh_seq = MeshSeq(self.time_partition, meshes)
+        mesh_seq = MeshSeq(meshes)
         got = re.sub("#[0-9]*", "?", str(mesh_seq))
         self.assertEqual(got, "['<Mesh ?>', '<Mesh ?>']")
 
-    def test_mesh_seq_time_interval_repr(self):
-        mesh_seq = MeshSeq(self.time_interval, [UnitSquareMesh(1, 1)])
+    def test_mesh_seq_repr(self):
+        mesh_seq = MeshSeq([UnitSquareMesh(1, 1)])
         expected = (
             "MeshSeq([Mesh(VectorElement("
             "FiniteElement('Lagrange', triangle, 1), dim=2), .*)])"
         )
         self.assertTrue(re.match(repr(mesh_seq), expected))
 
-    def test_mesh_seq_time_partition_2_repr(self):
+    def test_mesh_seq_2_repr(self):
         meshes = [
             UnitSquareMesh(1, 1, diagonal="left"),
             UnitSquareMesh(1, 1, diagonal="right"),
         ]
-        mesh_seq = MeshSeq(self.time_partition, meshes)
+        mesh_seq = MeshSeq(meshes)
         expected = (
             "MeshSeq(["
             "Mesh(VectorElement(FiniteElement('Lagrange', triangle, 1), dim=2), .*), "
@@ -177,13 +121,13 @@ class TestStringFormatting(BaseClasses.MeshSeqTestCase):
         )
         self.assertTrue(re.match(repr(mesh_seq), expected))
 
-    def test_mesh_seq_time_partition_3_repr(self):
+    def test_mesh_seq_3_repr(self):
         meshes = [
             UnitSquareMesh(1, 1, diagonal="left"),
             UnitSquareMesh(1, 1, diagonal="right"),
             UnitSquareMesh(1, 1, diagonal="left"),
         ]
-        mesh_seq = MeshSeq(self.time_partition, meshes)
+        mesh_seq = MeshSeq(meshes)
         expected = (
             "MeshSeq(["
             "Mesh(VectorElement(FiniteElement('Lagrange', triangle, 1), dim=2), .*), "
